按要求修改代码
1、修改loadmodel 和base.py 将模型初始参数可以导出，并可以导入，可以使用默认的model_export.py
2、使用spliter-cd 替换spliter.py
3、使用pres-ubmit 格式
4、移除stanet_predict教程
5、转换教程位置，并重命名，并保持原有风格，去掉命令行选项配置风格
